added antagonist anims and player sfx
added the latest animations for the antagonist to the ghost so when
approaching a raising object he will put his hand into the ground and
when the player falls too close to the ghost he puts his arm forward to
grab the player. restarting the level. I also added sfx to the player
character.